chore: Refactor the directory structure
- rename views to pages

- create assets/img and assets/placeholder directories

- create useInput, useSwitch, useToggle hooks

- shift configs in the config folder
