+++ client/src/config/Root.js
@@ -3,16 +3,11 @@ import { createStore, applyMiddleware, compose } from 'redux';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
 
-import reducers from './reducers';
+import reducers from '../store/reducers';
 
 export default ({ initialState = {}, children }) => {
-  const composeEnhancers =
-    window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
-  const store = createStore(
-    reducers,
-    initialState,
-    composeEnhancers(applyMiddleware(thunk))
-  );
+  const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
+  const store = createStore(reducers, initialState, composeEnhancers(applyMiddleware(thunk)));
 
   return <Provider store={store}>{children}</Provider>;
 };
