+++ client/src/index.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import App from './views/App';
-import Root from './Root';
+import App from './pages/App';
+import Root from './config/Root';
 
 ReactDOM.render(
   <Root>
     <App />
   </Root>,
-  document.querySelector('#root')
+  document.querySelector('#root'),
 );
